requirements.txt
blueprint==3.4.2
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
micawber==0.5.2
peewee==3.14.0
psycopg2-binary==2.8.6
Werkzeug==1.0.1
